+++ src/main/java/sn/axa/apiaxacnaas/entities/ContractGarantie.java
@@ -5,6 +5,7 @@ import lombok.*;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -27,12 +28,12 @@ public class ContractGarantie {
     @JoinColumn(name = "garantie_id", referencedColumnName = "id")
     private Garantie garantie;
 
-    private Double capitalMax; // 350 000
-    private Double capitalDejaVerse;
+    private BigDecimal capitalMax; // 350 000
+    private BigDecimal capitalDejaVerse;
 
     private Integer plafondNuitsParAn; // 30
     private Integer nuitsRestantes;
-    private Double montantParNuit; // 5 000
+    private BigDecimal montantParNuit; // 5 000
 
     @Column(updatable = false)
     @CreationTimestamp
